% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-card.R
\name{card_widget}
\alias{card_widget}
\title{Turn an htmlwidget into a web deployable, interactive Twitter card}
\usage{
card_widget(widget, output_dir = ulid::ulid_generate(),
  name_prefix = "wdgtcrd", preview_img,
  html_title = class(widget)[[1]], card_twitter_handle = "",
  card_title = html_title[[1]], card_description = "",
  card_image_url_prefix = "", card_player_url_prefix = "",
  card_player_width = 480, card_player_height = 480,
  background = "white", bundle_type = c("tgz", "zip"))
}
\arguments{
\item{widget}{an \code{htmlwidget}}

\item{output_dir}{the path to save the card-able widget to. If the directory does
not exist it will be created for you (recursively). The value will be
\code{\link[=path.expand]{path.expand()}}ed.}

\item{name_prefix}{the name-prefix for the widget's \code{.html} file and \code{preview_img} file.}

\item{preview_img}{the path to the local preview image for the card-able widget. This
file must exist and will be copied over to the deployable directory and renamed
(see \code{name_prefix} above). Follow the guidelines
\href{https://sproutsocial.com/insights/social-media-image-sizes-guide/#twitter}{here}
regarding image sizes.}

\item{html_title}{the title for the \code{htmlwidget} HTML file's \code{<title>} tag.}

\item{card_twitter_handle}{Your twitter handle \emph{including} the \code{@}.}

\item{card_title, card_description}{The title and description that will be displayed in the tweet}

\item{card_image_url_prefix}{Prefix URL for where you will be copying the preview image to.
Generally, this wil be the same as \code{card_player_url_prefix} but you can specify
another URL prefix if storing images on a separate server or separate directory.}

\item{card_player_url_prefix}{Prefix URL for where you will be copying the \code{htmlwidget}
HTML and supporting javascript libraries to. Generally, this wil be the same as
\code{card_image_url_prefix} but you can specify another URL prefix if storing images
on a separate server or separate directory.}

\item{card_player_width, card_player_height}{the width and height for the player window in-tweet.
These default to 480x480 and you should review the References section for
links to guidelines for Twitter preferred image sizes.}

\item{background}{\code{htmlwidget} background coloe. Defaults to \code{white}. Can be a hashed-prefixed
hex value (if so, will be converted to \code{rgba()} spec)}

\item{bundle_type}{either \code{tgz} for a gzip'd/tar archive or \code{zip} for a ZIP archive. The
directory named \code{name_prefix} will be placed into the archive.}
}
\value{
the \code{path.expand()}ed path to the \code{bundle_type}. The archive name will be \code{name_prefix}
plus the \code{bundle_type} extension.
}
\description{
Provide a preview image and a widget, plus Twitter Player card metadata and get pack
a packaged up, ready-to-deploy archive to deploy and used on Twitter.
}
\details{
You can use \href{https://cards-dev.twitter.com/validator}{Twitter's Validator} to ensure
your creation is usable before trying it in a tweet.
}
\note{
You can and should use \href{https://cards-dev.twitter.com/validator}{Twitter's Validator}
to ensure your creation is usable before trying it in a tweet.
}
\references{
\itemize{
\item \url{https://developer.twitter.com/en/docs/tweets/optimize-with-cards/overview/player-card.html}
\item \url{https://sproutsocial.com/insights/social-media-image-sizes-guide/#twitter}
\item \url{https://github.com/twitterdev/cards-player-samples}
}
}
