% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-preview.R
\name{gg_preview}
\alias{gg_preview}
\title{Generate a Twitter Player card preview image from a ggplot2 plot}
\usage{
gg_preview(gg, width = 350/72, height = 196/72, dpi = "retina")
}
\arguments{
\item{gg}{ggplot2 plot object}

\item{width, height}{width and height of the preview image. See References for guidelines}

\item{dpi}{see \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}}
}
\value{
path to the preview image tempfile
}
\description{
Takes arguments similar to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} and creates a temporary
plot file (png) to be used as input to \code{\link[=card_widget]{card_widget()}}.
}
\references{
\itemize{
\item \url{https://developer.twitter.com/en/docs/tweets/optimize-with-cards/overview/player-card.html}
\item \url{https://sproutsocial.com/insights/social-media-image-sizes-guide/#twitter}
}
}
