% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetcard-package.R
\docType{package}
\name{widgetcard}
\alias{widgetcard}
\alias{widgetcard-package}
\title{Tools to Enable Easier Content Embedding in Tweets}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/widgetcard}
\item BugReports: \url{https://gitlab.com/hrbrmstr/widgetcard/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
